campus-management-assistant/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
requests
beautifulsoup4
pdfplumber
google-generativeai
